Timer fragment display the first phase of the task list

+++ app/src/main/java/it/unipd/dei/esp2023/timer/Phase.kt
@@ -9,5 +9,5 @@ placeholders.
 data class Phase(
     val taskId: Long,
     val name: String,
-    var duration: Long
+    var duration: Int
 )
